#277 - extracting inner class GithubRepoStatFactory
